requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
psutil==5.9.8
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
scikit-learn==1.3.2
apscheduler==3.10.4
websockets==12.0
